hook: Fix error message when reference-transaction hook fails
The error message when executing the reference-transaction hook logic
via the manager presumes that we're about to vote, but we should rather
treat it as a blackbox and only say that the reference-transaction hook
has failed. This also fixes the weird error message we potentially get
here: "error voting on transaction: error voting on transaction:
transaction was aborted".