git: Move GITALY_GIT_BIN_PATH into the hooks payload
We've recently grown another new environment variable which is being
passed to hooks in 348e6869e (Make git-upload-pack use gitaly-hooks for
pack-objects, 2021-01-21). Nowadays, to help discoverability and
immediately see what kind of information any hook expects, such
information is instead put into the HooksPayload structure.

This commit thus adds a new field `GitPath` into the struct. Given that
the code has been deployed already, we need to provide a fallback which
extracts the value from the old GITALY_GIT_BIN_PATH environment
variable.

To demonstrate that the fallback works as expected, this commit already
modifies the single point of usage of that envvar to use the payload's
git path instead, without inject it into the payload directly. Given
that all tests pass, we thus know that the fallback works as expected
and that it uses the envvar's value.