uploadpack: Improve test coverage for pack-objects hook
We've recently grown support for setting up the pack-objects hook. The
test coverage is kind of lacking though: we do have several tests which
now set the accompanying feature flag, but we don't assert that the hook
is being executed in the first place. Furthermore, only the smarthttp
service is tested right now, with no tests for the ssh service.

This commit improves coverage by adding two tests which assert that
the hook is executed correctly and adding additional testcases which set
up the feature flag for the ssh service.