git: Relax hook checks to allow hooks for non-modifying commands
Setup of hooks is currently only allowed for commands which modify any
references. This stems from the fact that all existing hooks were only
ever executed when any refs changed, so this check was added as a safety
measure to not setup hooks if not required.

We've recently grown support for another hook though with the
pack-objects hook. This one is different than the existing hooks, mainly
that it's executed for a read-only command. So in this context, it
doesn't make a lot of sense to keep the restiction.

Let's relax the restriction to allow configuration of hooks for
read-only commands. The requirement to set up hooks for modifying
commands is left as-is.