git: Use `configureHooks()` to set up pack-objects hooks
The pack-objects hook sets up the command's hook configuration manually,
but most of this setup is duplicated with what `configureHooks()` would
do. This has been implemented such because previous to the preceding
commit, we didn't allow to configure hooks for read-only commands. Now
that we do, let's refactor the code to also use `configureHooks()`.

Note that this brings with it two changes: first, we now get the
safeguard to not allow configuration of hooks multiple times for free.
Second, we now potentially inject transaction information into the
HookPayload. Given that the hook is read-only, this information wouldn't
ever be used, so it's harmless.