add default replication factor to virtual storage's configuration
Adds a configuration option for the default replication factor to
use for new repositories in a given virtual storage. This allows
Praefect to configure new repositories directly with the desired
replication factor instead of the administrator having to manually
set the replication factor of a repository after it has been created
already.