git: Implement NewReferenceFromBranchName function
Quite often, one needs to convert from a unqualified branch name to a
fully qualified branch name, e.g. from "master" -> "refs/heads/master".
To do this, we've currently got `NewBranchReferenceName()`, which takes
a branch name and converts it to a fully-qualified reference.

This method is actively dangerous though: in order to provide
compatibility with legacy code, it was initially implemented in such a
way that it will treat a branch name which starts with "refs/heads/" as
an already-fully-qualified reference. It is, but assuming that callers
get calling semantics correct, the original intention was to get a
branch which is called "refs/heads/master" with its fully qualified
reference "refs/heads/refs/heads/master". It's an awkward reference
name, but who are we to judge? The result is confusion and potentially
we modify or read the wrong reference.

Let's thus introduce a new function `NewReferenceNameFromBranchName()`
which will always prepend "refs/heads/" no matter what.