ref: Fix FindRef RPC confusing reference names
The `FindRef()` RPC accepts as parameter an unqualified branch name. We
thus need to convert it to a fully qualified branch name in order to not
cause ambiguous reference lookups in the RPC's implementation.

To convert the branch name, we currently use `NewBranchReferenceName()`,
which has dangerous semantics as it will treat a branch name starting
with "refs/heads/" as an already-qualified branch. As a result, we'd now
start to look up the "master" branch instead of the "refs/heads/master".

Let's fix this bug by converting to `NewReferenceFromBranchName()`. Note
that this changes semantics of the RPC to only support unqualified
branch names. There is no way around breakage though, as otherwise we
wouldn't be able to cope with ambiguous references.