operations: Fix UpdateSubmodule RPC confusing reference names
The `UpdateSubmodule()` RPC accepts as parameter an unqualified branch
name. We thus need to convert it to a fully qualified branch name in
order to not cause ambiguous reference lookups in the RPC's
implementation.

To convert the branch name, we currently use `NewBranchReferenceName()`,
which has dangerous semantics as it will treat a branch name starting
with "refs/heads/" as an already-qualified branch. As a result, we'd now
start to lookup the "master" branch instead of the "refs/heads/master"
branch. This is fixed by converting to `NewReferenceFromBranchName()`.

Naturally, our legacy Ruby implementation has similar bugs. First, it
uses `rev_parse()` on the unqualified branch name, which may potentially
look up a tag instead of a branch if both have the same name. And
second, it uses the `branches` collection to look up the branch. But
this function provided by Rugged has the same dangerous semantics as
`NewBranchReferenceName()`, as it will treat a fully-qualified reference
not as the branch name, but as the branch reference. Those bugs are
fixed by passing a fully-qualified reference to `rev_parse()` and using
`ref()` with a fully-qualified reference instead of the `branches`
collection.