hook: Dependency-inject transaction manager
The transaction.Manager is currently instantiated by the hook.Manager
directly, which was done in order to make splitting up transaction and
hook logic focussed on the task without doing global changes in the
preceding commit. We do want to dependency inject it from external
though given that hooks aren't going to be the single location where
they're used.

So let's do so and adjust all sites where the hook manager is
instantiated to also pass a transaction manager.