git: Move setup of uploadpack.packObjectsHook config entry
The `uploadpack.packObjectsHook` config entry is currently being set up
manually both in the smarthttp and the ssh service. It's thus easy to
get out of sync if it ever needs to change and easy to forget in case we
grow additional callsites where the pack-objects hook needs to be
injected. We've thus started to inject such paths via the hook options
directly.

Let's do the same for `WithPackObjectsHookEnv()` and move setup of the
config entry in there.