intercept RepositoryExists calls in Praefect
GitLab calls RepositoryExists prior to creating a repository. Since
Praefect has no records of the repository before it is created,
PerRepositoryRouter's accessor routing fails as it doesn't know
where to send the requests without the database records.

This commit fixes the problem by intercepting RepositoryExists in
Praefect and checking from the database whether the repository
exists or not. Praefect then responds with the information without
involving the coordinator or any Gitalys.

This only affects the repository specific primary stack as the SQL
elector based stack just redirects the call to the virtual storage's
primary if no records of a repository exist.