Rename RevertConflictError to HasConflictsError
Prior to this change, the `UserRevert` command was the only command to
be handling a conflict error. In one of the subsequent commits
`UserCherryPick` will also be rewritten into Golang, and it also has to
handle a conflict error.

This change renames `git2go.RevertConflictError` to
`git2go.HasConflictsError` so the name is more generic and it makes
sense to be reused by `UserCherryPick`.