Makefile: Build libgit2 as part of prepare-tests
Given that libgit2 is required to run our Go test suite, it is being
built as part of our `test-go` Makefile target. Given that we have a
`prepare-tests` target, this is a bit awkward as a developer may righlty
assume that after having run `prepare-tests`, that he is able to now run
all of our tests. But the dev will be quick to note that all targets
which depend on libgit2 will fail now.

Let's improve this by moving the `libgit2` dependency from `test-go` to
`prepare-tests`.