transactions: Report voting hash on failed vote
When voting on a transaction fails because quorum wasn't reached, then
we print an error message detailing how many votes we got on the node's
vote. We're still up to a guessing game though which node has agreed or
disagreed on what exact vote, as there is no information about what the
node voted on.

Fix this by including the hash the node voted on in the error message.