coordinator: Log when RPC creates no subtransactions
When an transactional RPC creates no subtransactions, then we do not
know why that is: it either could be that the RPC is not aware of
transactions and thus wasn't able to correctly use transactions, or that
some git command didn't use hooks because of a programmatic error, or
that the RPC simply didn't change anything. Because we cannot know what
happened, we need to assume the worst: that the repository is now out of
sync across the replicas. As a logical result, we thus always create a
replication job in case no transaction happened to be on the safe side.

This replication job is wasteful, as in most cases we probably did
perform the changes correctly on all nodes. But because we don't log
this, it's hard to observe which RPCs cause this, and thus also hard to
fix. So let's improve observability by creating a log entry.