add repository creation routing
NodeManagerRouter routes repository creations as normal mutator calls.
This works as it defaults to routing to the virtual storage's primary
when no database records of a repository exist.

This does not work for the PerRepositoryRouter as there is no concept
of virtual storage's primary. Due to this, repository creations need
to be routed in a special manner.

This commit adds support for routing repository creations with
PerRepositoryRouter. It picks a random healthy node to act as the
repository's primary node and sets other nodes as replication targets.
This behavior matches the existing behavior in NodeManagerRouter, except
for picking a random node to act as the primary instead of the virtual
storage's primary.

The functionality is not yet wired to the rest of the code.