handle repository creations separately in coordinator
This commit wires up repository creation routing to Praefect's
coordinator. If the RPC creates a repository, the routing is handled
different manner from other repository scoped mutators.

This fixes repository creation when repository specific primaries are
enabled. Prior to this, repository creation did not work as
PerRepositoryRouter needs to know where to route the mutator request.
As there were no record of the repository, the router didn't know
where to route the request. The router now picks a random healthy node
to act as the primary and sets the other nodes as replication targets.
It then stores the primary in the database in the request finalizers.