transactions: Move signalling logic into separate function
The signalling logic is quite complex, as there are multiple different
conditions which require us to signal. It's still part of `vote()`,
which has forced us to have multiple exit paths in that function which
signal. This fact is easy to miss and confusing, so this commit pulls
out the logic to determine whether we need to signal or not into a
separate function. This also improves our ability to test the logic,
which we'll make use of in a subsequent commit.