transactions: Fix race between updating voter state and getting state
The voter state is being updated in three different locations:

    - `stop()` will update voters to `VoteStopped` in case it's yet
      undecided.

    - `cancel()` will update voters to `VoteCanceled` in case it's yet
      undecided.

    - `collectVotes()` will update voters either to `VoteFailed` or
      `VoteCommitted`, depending on whether it's part of a potential
      quorum or not.

While the first two ones use a write-lock, the last one only uses a
read-lock. As a result, it can race with `getState()`, which will read
the state of each node which `collectVotes()` may be about to update.

Fixing this isn't quite trivial. We could just switch from a read-write
mutex to a normal mutex, but this feels conceptually untidy. Instead,
this commit refactors the code to perform vote counting as part of
`vote()`: when either all votes were cast or when quorum was reached,
we'll iterate over all undecided nodes and update their state to either
`VoteFailed` or `VoteCommitted`. As a result, we shouldn't be able to
observe any undecided nodes in `collectVotes()` anymore and thus don't
have to update any state anymore, making it purely read-only.