transactions: Merge finalization channels of subtransactions
We have multiple ways of finalizing a subtransaction either by canceling
it, stopping it or when enough votes have been collected to either reach
quorum or fail the vote. These three conditions currently use three
different channels to signal that either of them has happened. The
original intent was that we needed to make sure that no channel is
closed twice, which would've resulted in a panic. But it's still kind of
awkward to handle.

This commit refactors the code to instead use a single `doneCh` channel
as well as a boolean flag indicating whether the subtransaction has been
finalized already, which stops us from closing the channel multiple
times.

By merging the channels, we loose the ability to determine whether a
transaction was canceled or stopped based on which channel was closed.
But it turns out this wasn't needed in the first place: when canceling
or stopping a transaction, we set the state of any undecided voter to
`VoteCanceled` or `VoteStopped` already, and we do check those states
and return `ErrTransactionCanceled` respectively `ErrTransactionStopped`
already. So these checks were redundant.