transactions: Only allow voting on undecided voters
When a transaction has been canceled or stopped, then we should not
allow any node to cast additional votes. While we'd have returned an
error when collecting votes at a later point, we didn't prohibit the
node from casting its vote in the first place. The result is that vote
counts are off in the case where a stopped or canceled node cast its
vote, leading to a potentially inconsistent representation of reality.

Fix this issue by disallowing any vote of a voter which is not in
VoteUndecided state. This is even stricter than only disallowing casting
votes for stopped/canceled voters, but given that we do not allow
casting votes multiple times it just makes the invariant clearer.