transactions: Simplify signalling logic
Now that we have a flag indicating whether a transaction was finalized
already, we can greatly simplify our signalling logic. Previously, we
needed to make sure that we didn't accidentally close the `doneCh`
channel twice. Given that this information wasn't readily available, we
had to reevaluate every time whether any node would've already notified
voters based on the previous state of vote counts.

Now that we have a simple flag available, we can simplify the signalling
logic quite a bit: if it's set already, we know that the transaction was
finalized already and thus we mustn't notify voters. Otherwise, we can
simply check whether any node has crossed the threshold or alternatively
if all votes have been cast.