transactions: Allow casting all-zero votes
The voting logic currently treats the all-zero vote as a placeholder for
no vote having been cast by a voter. And this has been fine until
recently where we only ever cast votes on the hashed contents of
reference updates: hashing something to all-zeroes is unlikely enough.

This has changed with 77c0166c3 (repository: Use transactions when
writing gitattributes, 2021-01-28). If using transactions for writing
gitattributes, we'd cast a vote of all-zeroes when deleting the file.
Praefect would've considered this as no vote though and pretended that
the primary just didn't vote.

Fix this by converting the vote into a pointer: if it's `nil`, no vote
was cast. Otherwise, it represents a valid vote. This allows us to
remove the ambiguous meaning of all-zero votes.