evaluate repo path's symlinks in UserUpdateBranch tests
macOS' default temporary directory is under '/var/...'. '/var' is a
symlink to '/private/var'. UserUpdateBranch tests compare the full path
with symlinks evaluated to a path that does not have the symlinks expanded,
causing the test to fail on macOS. This commit evaluates the symlinks prior
to comparing the two paths to fix this.