log errors in User{Create,Update,Delete}Branch RPCs
UserCreateBranch, UserUpdateBranch and UserDeleteBranch do not log
errors they are dropping in favor of alternative error messages.
Some errors are also packed in to RPC responses which. This causes
the errors to look like successful responses which do not get logged
by Gitaly's gRPC logging interceptor. These behaviors can make
debugging difficult as it's not clear what was the original cause of
the error. This commit addresses the situation by logging the swallowed
errors.