support storing assignments on repository creation
When variable replication factor is enabled, the repository's host
nodes should be selected when the repository is being created. The
selected assignments should then be persisted so they can be used
for future routing decisions. This commit adds support for storing
assignments in RepositoryStore's CreateRepository.

The functionality is behind a boolean switch as assignments are not
supported when repository specific primaries are not used. Separate
switch from storing primary is used in order to support repository
specific primaries without variable replication factor. This allows
for independently rolling out the features to production later.