support default replication factor in coordinator
This commit adds support for default replication factor in
Praefect's coordinator and hooks up the configuration option with
the rest of the code.

When a default replication factor is configured for a virtual storage
and repository specific primaries are enabled, the coordinator stores
the nodes participating in the repository's creation as the assigned
nodes.