log: Add missing newline to fatal hook error messages
The gitaly-hooks executable uses a special HookLogger because it
typically isn't allowed to write to either its stdout nor its stderr as
both are potentially used by git itself. It thus writes into a logfile
instead as we do not yet have the infrastructure yet to directly relay
log messages to the Gitaly server.

The exception for this is when a fatal error occurs though: besides
logging to the file, gitaly-hooks also prints a single line "error
executing git hook". This is then relayed to the server via git's normal
stderr and thus shows up in our log messages.

The log message is missing a trailing newline though. As a result,
generated error messages look for example like:

    error executing git hookfatal: ref updates aborted by hook

This is obviously wrong, so let's fix it by using `fmt.Println()`
instead of `fmt.Printf()`.