hook: Log errors when stopping transactions fails
When we have to stop transactions because some primary-specific logic
has failed, then we currently completely ignore any errors which happen
when stopping the transaction itself. We can't really help this given
that the initial error is the actual root cause and thus the error we
want to return. But we can at least improve the code by logging those
issues.