localrepo: Expose functions to execute commands
When initially implementing the localrepo abstractions, the plan had
been to migrate ad-hoc executions of git commands into more abstract
functions in order to not have to distribute the knowledge about how to
execute certain git commands throughout our codebase. That's also why the
localrepo does not expose any way to execute git commands on that repo
diretcly.

The migration isn't really happening though, or only with a very slow
pace. So let's expose two ways to run git commands in the context of a
localrepo with `Exec()` and `ExecAndWait()`. These take away the need to
instantiate git command factories everywhere, and especially the latter
takes away the pain point that one always needs to do two error checks
when executing a git command.

The previous `command()` function will be removed in a subsequent commit
when all its users have been migrated.