log: Fix use of protobuf repo
The `internal/git/log` package accepts a `*gitalypb.Repository` instead
of a `repository.GitRepo` in some places. Use of the protobuf is
discouraged wherever it can be avoided, so let's convert the interfaces
to instead accept a `repository.GitRepo`. This also allows us to pass in
a `localrepo.Repo`, which implements `repository.GitRepo`.