uploadpack: Fix racy tests for pack-objects hook
The pack-objects hook tests are racy. While I cannot reproduce the
flakiness on my machine, it seems like the transport is closing before
the error message which we inject via hook is sent.

Let's rewrite the tests such that they do not produce an error, but
instead simply intercept the git-pack-objects(1) invocation to write a
given output file and then afterwards execute the real git binary. This
should remove the racy error case altogether as there are no errors
anymore.