Refactoring of the GrpcErrorHasMessage
GrpcErrorHasMessage returns a boolean flag and not always
used in a right way. We change it's signature to exclude
its misuse. Refactored implementation showed a bug in one
of the tests that was fixed: without sending response back
to the stream the client returns io.EOF error for each request.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/2699