Refactoring of the test dependent on runPraefectServerWithGitaly
Helper function runPraefectServerWithGitaly is hard to change
and configure. It changes the state of the passed arguments
internally which is not obvious and requires developers to spend
a lot more time to work with it. We can't use a new config builder
as there are still dependencies on the global config.Config variable
in the implementation those tests verify.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/2699