Replace usage of git.NewCommandWithoutRepo with factory method
To break dependency on the gloabl config.Config variable we
introduced git.CommandFactory abstraction that hides all the
dependencies internally and can be mocked if needed. This
gives us more flexibility in testing and makes the code less
buggy by removing dependency on the global state. This change
replaces all usages of the NewCommandWithoutRepo function with
NewWithoutRepo method of the CommandFactory abstraction and
git.CommandFactory changed to use its private newCommand method
to run the command instead of NewCommandWithoutRepo.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/2699