git: Remove readonly flag
The `scReadonly` flag for git commands has been used in order to
determine whether a given command needs the reference-transaction or
not. It is used in the same context as the `scNoRefUpdates` flag though
and doesn't have any additional meaning in other contexts.

Remove the flag in favor of `scNoRefUpdates`.