hook: Use explicit stat to avoid manually resolving symlinks
When creating a custom hooks executor, we need to scan through the
custom hooks directories in order to collect all hooks. We need to be
careful to not try and execute anything which is not an executable file.
Traditionally, this was done by simply checking the directory entry's
file mode to see whether it has the executable bit, but this fails both
for directories and for symlinks, both of which typically have the bit
set. Skipping directories is easy enough, but for symlinks we need to
check the symlink target's mode.

To evaluate the target, we're currently going the quite roundabout way
of first evaluating the symlink and then doing an lstat(3P) on the
resolved target. This is more complex than it needs to be, given that we
can just stat(3P) the symlink directly as this syscall will resolve any
symbolic references for us.

Let's simplify the code by using stat(3P) instead of manually resolving
and using lstat(3P). While at it, this commit moves all of those checks
into the `validHook()` function such that we can reuse it in other
places, too. The downside is that we now always perform an additional
stat(3P) call to verify whether non-symlink are valid hooks. But this
shouldn't really be noticeable performance-wise.