hook: Unify checks whether a hook is valid
While we always use `isValidHook()` to check hooks in custom-hook
directories, we use an open-coded variant of this check to check whether
the per-project custom hooks file is valid or not. But instead of using
`unix.Access()`, it directly checks for the user-executable bit of the
file, which may not be an accurate representation of whether we're
allowed to execute the file or not.

Fix this by using `isValidHook()` here, too.