testhelper: Raise default log level
When configuring our tests, the default log level gets set to "info"
level. This causes us to print a lot of irrelevant logs to the console
which ultimately don't really help to see what's going on.

Let's raise the error to "panic". While this is quite high, all other
levels get hit all the time as we're also testing error cases.

Given that a testcase in `internal/tempdir` depends on reading
info-level log entries, this commit also fixes that testcase to
explicitly set the log level. Note that because the global logger cannot
easily be restored, the test just keeps the logger's changed
configuration after the test.