grpc-proxy: Do not print error messages to stderr
The grpc-proxy proxy handler tests set up a separate logger for GRPC log
messages, but we're already setting one up as part of the testhelper
configuration. This leads to a spew of log messages when those tests get
executed.

Fix this by just dropping the separate logger in favor of the testhelper
one. While that one is configured to only print messages with the
"panic" log level by default, this is rather considered a feature and
not a bug.