grpc-proxy: Remove explicit logging
The grpc-proxy proxy tests explicitly print logs both when we're
starting GRPC server and proxy as well as when handling requests. This
doesn't help in normal test runs, and when an error is hit in those
tests then the log messages probably wouldn't help either.

Let's drop those messages altogether to reduce the amount of distracting
messages.