testhelper: Decrease logging verbosity for Praefect
When running our test suite with Praefect proxying enabled, then we
start a separate Praefect server whenever we set up a Gitaly test
server. This Praefect instance is configured with a default log level,
which is "info", and as a result the tests become extremely verbose.

Fix this by explicitly passing a logging configuration to the Praefect
server.