transactions: Increase log level for committed transactions
In order to fully understand the timeline of a transaction, we not only
need to know about failed transactions, but also about successful ones.
The "transaction completed" log entry in Praefect's transaction manager
only logs with a "debug" log-level though, which by default will not get
get logged at all.

Improve this situation by raising the log level to "info".