gitaly-hooks: Check for valid hooks before dialing Gitaly
When gitaly-hooks is executed with an unknown command, we'll first try
to dial Gitaly and only afterwards fail with an error because we cannot
handle the requested hook. This is inefficient, but has been hard to fix
previous to the preceding commit which pulled out the hook logic into
its own function.

Now that we do have those functions, this commit further refactors the
code to have a map of hooks by their respective subcommand. This allows
us to do an easy lookup of the hook function which we can now move
before the dialling logic. While this doesn't fix any known bug, it
feels cleaner and allows us to easily extend hook commands to also carry
their `git.Hook` type in a subsequent commit.