gitaly-hooks: Honor requested hooks
The HooksPayload has grown a new bitmap of requested hooks: only if the
bit corresponding to the about-to-be-executed hook is set shall we
actually execute that logic. But gitaly-hooks doesn't yet pay attention
to this flag, which this commit now fixes.

Note that we still do not inject the requested hooks anywhere. This is
done on purpose to demonstrate that the fallback of simply treating any
hook as requested works correctly to not break zero-downtime upgrades in
case the new gitaly-hooks binary is installed already while the old
server doesn't yet know about requested hooks.