helper: Remove `ProtoRepoFromRepo()`
Now that the reference-transaction hook can be set up with a
`repository.GitRepo` instead of with a `gitalypb.Repository`, this
commit removes the `helper.ProtoRepoFromRepo()` function and instead
just passes down the interface directly. This gets rid of a potentially
dangerous function which could have lead us to hand down
partially-populated protos to unsuspecting callees.