gitaly: Set up requested hooks bitmap
With the infrastructure being in place to support executing requested
hooks only, this commit converts all callsites of `NewHooksPayload()` to
also pass the bitmap of hooks they want to execute. This finally fixes
an issue where we could've accidentally executed an unwanted hook only
because we set up `core.hooksPath` for a git command.