objectpool: Introduce localrepo
The localrepo abstraction has only been used in contexts of a "real"
repository until now. But given that it requires project-independent
information about repos only, it depends on the `repository.GitRepo`
interface instead of the repository protobuf. As such, there is no
reason it shouldn't be used for object pools, too.

This commit introduces a new `poolRepo` struct member to the
`ObjectPool` structure. This new fields isn't used yet, but we'll do so
in subsequent commits.