localrepo: Implement `remote.Exists()`
Both the `git/localrepo` and `git/remote` packages implement support for
handling remotes with part of the functionality existing in both
packages. The localrepo is the new any shiny interface though and much
more complete than what the remote package provides. The only exception
is the `remote.Exists()` function which checks whether a remote exists
or not, which doesn't yet exist in the localrepo package.

This commit ports it over such that the localrepo package can serve all
usecases the remote package can. The new function isn't yet used, but
all callers will be converted in the next commit.