Removal of git.NewCommand
After long rounds of refactoring we finally can remove
git.NewCommand function.
The change also includes removal of the TODO left because
of the NewCommand usage with the global config.Config variable.
And as a final step the doc is updated and now mentions usage
of the git.CommandFactory interface instead of git.NewCommand().

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/2699