localrepo: Limit for-each-ref when getting a single ref only
The `localrepo.GetReference()` function is only ever interested in a
single reference, but we call git-for-each-ref(1) without any limit. In
most cases this shouldn't matter as any exact matches would typically
only return a single reference anyway. But if we are for example given a
prefix, then we would first list all references of that prefix only to
fail afterwards because we didn't find the expected reference.

Fix this issue by limiting git-for-each-ref(1) to print a single
reference, only.