editorconfig: Set line length for Go code
The editorconfig is currently lacking a line length configuration for Go
code. This is probably by design given that we do not have an explicit
limit on line length in our style guide. But it's still helpful to have
a reasonable limit set, even if we're not going to enforce it.

Let's configure a line limit of 100 characters. This is not a new
requirement but should rather be seen as a recommendation and help for
users who use editorconfig.